"Added authentication and authorization features, including Django Rest Framework's token authentication and allauth, and updated models, serializers, and views accordingly."

+++ ecomapi/ouedkniss/models.py
@@ -26,7 +26,7 @@ class Order (models.Model):
 class Paiement(models.Model):
     order = models.OneToOneField(Order, on_delete=models.CASCADE)
     client = models.ForeignKey(Client, on_delete=models.CASCADE)
-    amount = models.IntegerField(max_digits=10, decimal_places=2)
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
     payment_method = models.CharField(default="online", max_length=20)
     status = models.CharField( default="pending" ,max_length=20,)
     transaction_date = models.DateTimeField(auto_now_add=True)
